requirements.txt
azure-functions
azure-cli
pandas